:recycle: refactor(urlmapping): change the clicks type to long instead of integer
The UrlMapping entity now uses Long as its primary identifier, replacing the Integer type.

+++ src/main/java/com/github/machadojoaovictor/link_shortener/entity/UrlMapping.java
@@ -32,11 +32,11 @@ public class UrlMapping {
 
     private LocalDateTime expiresAt;
 
-    private Integer maxClicks;
+    private Long maxClicks;
 
     @Builder.Default
     @Column(nullable = false)
-    private Integer clicks = 0;
+    private Long clicks = 0L;
 
     @Builder.Default
     @Enumerated(EnumType.STRING)
